plugins/velociraptor_legacy/requirements.txt
# List third-party dependencies here, separated by newlines.
# All dependencies must be version-pinned, eg. requests==1.2.0
# See: https://pip.pypa.io/en/stable/user_guide/#requirements-files
paramiko==3.3.1
grpcio==1.59.2
grpcio.tools==1.59.2
pyvelociraptor==0.1.8
